requirements.txt
python-telegram-bot==21.7
anthropic==0.42.0
python-dotenv==1.0.1
